pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deckgen"
version = "1.0.0"
description = "Generate beautiful HTML presentations from context"
authors = [{name = "cpark4x"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
deckgen = "deckgen.cli:cli"

[project.urls]
Homepage = "https://github.com/cpark4x/deckgen"
Repository = "https://github.com/cpark4x/deckgen"
